createViewModelComponent code review changes
